- Global configuration feature - Minor code improvements

+++ database/migrations/2019_02_07_093051_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('sku');
             $table->boolean('status');
             $table->double('basePrice');
-            $table->double('specialPrice')->nullable();
+            $table->float('specialPrice')->nullable();
             $table->string('image')->nullable();
             $table->longText('description');
             $table->timestamps();
